improved automated tabulation

+++ src/analysis/CoreQuery.m
@@ -6,7 +6,7 @@ classdef CoreQuery < handle
     % them.
     
     properties ( SetAccess = private )
-        count(1,1) uint32
+        count(1,1) double
         label_array(:,:,:) uint32
     end
     
@@ -41,7 +41,7 @@ classdef CoreQuery < handle
         end
         
         function value = get.count( obj )
-            value = uint32( obj.cc.NumObjects );
+            value = obj.cc.NumObjects;
         end
         
         function value = get.label_array( obj )
