+++ src/analysis/PartingLineQuery.m
@@ -5,7 +5,7 @@ classdef PartingLineQuery < handle
     end
     
     properties ( SetAccess = private, Dependent )
-        count(1,1) uint32
+        count(1,1) double
         label_array(:,:,:) uint32
         binary_array(:,:,:) logical
         draw(1,1) double % mesh units
@@ -51,7 +51,7 @@ classdef PartingLineQuery < handle
         end
         
         function value = get.count( obj )
-            value = uint32( obj.cc.NumObjects );
+            value = obj.cc.NumObjects;
         end
         
         function value = get.label_array( obj )
