+++ src/analysis/ProjectedPerimeterQuery.m
@@ -6,7 +6,7 @@ classdef ProjectedPerimeterQuery < handle
     end
     
     properties ( SetAccess = private, Dependent )
-        count(1,1) uint32
+        count(1,1) double
         label_array(:,:) uint32
         binary_array(:,:) logical
     end
@@ -29,7 +29,7 @@ classdef ProjectedPerimeterQuery < handle
         end
         
         function value = get.count( obj )
-            value = uint32( obj.cc.NumObjects );
+            value = obj.cc.NumObjects;
         end
         
         function value = get.label_array( obj )
